T4, and DataFormat content
-Varsr woerk with DataFormat
-Format holder soon going to become obsolete
-Spiknt now has aoe heal, new setting for advanced effects
-Instant BOuncinb Bullets display effect both at ending point and at owner pos so they can see what path the bullet took
-Added spinascene, a t4 forgotten unit which sprays it's voidic shards everywhere, and uses a compination of both normal and instant bouncing bullets in it's attacks. Can be used extremely effectively against crowds or small buildings.
-Moar bullats
-DysfunctionalMonolith onw blinks whilst game is paused

+++ core/src/rusting/entities/bullet/BounceBulletType.java
@@ -20,7 +20,7 @@ public class BounceBulletType extends ConsBulletType {
     //How thick trail is
     public float trailWidth = 1;
 
-    public BounceBulletType(int speed, int damage, String sprite) {
+    public BounceBulletType(float speed, float damage, String sprite) {
         super(speed, damage, sprite);
         this.sprite = sprite;
         this.speed = speed;
